bugfixes, find unused mods

+++ src/import_ls.py
@@ -13,6 +13,7 @@ import logging as log
 import pysed
 import hashlib
 import pathlib
+
 existing_mods = {}
 
 def importAllMods(path, rem = False):
@@ -32,6 +33,7 @@ def importAllMods(path, rem = False):
 
 def getMods(path):
 	mods = []
+	all_mods = os.listdir(se.getSettings('all_mods_path'))
 	try:
 		files = os.listdir(path)
 		for i in files:
@@ -39,6 +41,15 @@ def getMods(path):
 				with zipfile.ZipFile(path + os.sep + i) as z:
 					try:
 						moddesc = ET.fromstring(z.read('modDesc.xml').decode('utf8').strip())
+						version = moddesc.find('version')
+						descV = int(moddesc.attrib['descVersion'])
+						f_name = 'fsl_' + version.text + '!' + i
+						if f_name in all_mods:
+							moddesc = None
+						if se.vers == 'fs19' and int(descV / 10) >= 6:
+							moddesc = None
+						if se.vers == 'fs22' and int(descV / 10) < 6:
+							moddesc = None
 					except FileNotFoundError:
 						moddesc = None
 						pass
@@ -49,6 +60,8 @@ def getMods(path):
 						mods.append(i)
 	except FileNotFoundError:
 		pass
+	if not mods:
+		sg.popup_ok(tr.getTrans('no_mod_found'), title = '')
 	return mods
 
 
@@ -94,7 +107,7 @@ def updateSGS(sgs, mod):
 		# TODO map update
 	return
 
-def importMods(path, mods, updateSGs):
+def importMods(path, mods, updateSGs, rem = False):
 	all_mods = se.getSettings('all_mods_path')
 	for i in mods:
 		with zipfile.ZipFile(path + os.sep + i) as z:
@@ -118,7 +131,7 @@ def importMods(path, mods, updateSGs):
 				else:
 					mod_type = 'mod'
 				if d == None:
-					TinyDB(all_mods + os.sep + 'mods_db.json').insert({'name': name.text, 'mod_type': mod_type, 'lang': mod_lang, 'files': {f: f_hash}})
+					TinyDB(all_mods + os.sep + 'mods_db.json').insert({'name': name.text, 'mod_type': mod_type, 'lang': mod_lang, 'files': {new_name: f_hash}})
 				else:
 					d['files'][new_name] = f_hash
 					TinyDB(all_mods + os.sep + 'mods_db.json').update({'files': d['files']}, doc_ids = [d.doc_id])
@@ -141,11 +154,32 @@ def importMods(path, mods, updateSGs):
 								updateSGS(values['-SGS-'], new_name)
 							window.close()
 							break
+		if sg.popup_yes_no(tr.getTrans('rem_source').format(path + os.sep + i), title = tr.getTrans('remove_title'), location = (50, 50)) == 'Yes' or rem:
+			os.remove(path + os.sep +i)
 
 def removeMods(mods):
 	all_mods = se.getSettings('all_mods_path')
+	datasets = TinyDB(se.games_json).all()
 	for i, val in enumerate(mods):
-		os.remove(all_mods + os.sep + existing_mods[val])
+		unused = True
+		for i in range(len(datasets)):
+			if existing_mods[val] in list(datasets[i]['mods'].values()) or existing_mods[val] in datasets[i]['map']:
+				sg.popup_ok(tr.getTrans('cant_rem_mod').format(existing_mods[val], datasets[i]['name']), title = tr.getTrans('error'))
+				unused = False
+				break
+		if unused:
+			with zipfile.ZipFile(all_mods + os.sep + existing_mods[val]) as z:
+				moddesc = ET.fromstring(z.read('modDesc.xml').decode('utf8').strip())
+				icon = moddesc.find('iconFilename').text
+				for l in se.getLangs():
+					name = moddesc.find('title/' + l)
+					mod_lang = l
+					if name != None:
+						break
+				d = TinyDB(all_mods + os.sep + 'mods_db.json').get(Query().name == name.text)
+				d['files'].pop(existing_mods[val])
+				TinyDB(all_mods + os.sep + 'mods_db.json').update({'files': d['files']}, doc_ids = [d.doc_id])
+			os.remove(all_mods + os.sep + existing_mods[val])
 
 def getAllMods():
 	global existing_mods
@@ -154,7 +188,24 @@ def getAllMods():
 	existing_mods.update(m[1])
 	for i in list(se.getInternalMaps().keys()):
 		del existing_mods[i]
-	return list(existing_mods.keys())
+	return list(sorted(existing_mods.keys()))
+
+def markUnusedMods(window):
+	datasets = TinyDB(se.games_json).all() 
+	maps, mods = ga.getMods(False)
+	mods.update(maps)
+	used = []
+	for dataset in datasets:
+		used = used + list(dataset['mods'].values())
+		used.append(dataset['map'])
+	print(used)
+	used = list(dict.fromkeys(used))
+	all_files = list(mods.values())
+	unused = []
+	for mod in all_files:
+		if not mod in used:
+			unused.append(list(mods.keys())[list(mods.values()).index(mod)])
+	window.Element('-MODS_INST-').SetValue(unused)
 
 def guiImportMods(updateSGs = True):
 	layout =    [	[sg.Text(tr.getTrans('get_mod_path'))],
@@ -165,7 +216,7 @@ def guiImportMods(updateSGs = True):
 					[sg.Button(tr.getTrans('import'), key = '-IMPORT-', size = (96, 1))],
 					[sg.Text(tr.getTrans('existing_mods'))],
 					[sg.Listbox(getAllMods(),  key = '-MODS_INST-', size = (108, 10), select_mode = 'extended')],
-					[sg.Button(tr.getTrans('remove'), key = '-REMOVE-', size = (96, 1))],
+					[sg.Button(tr.getTrans('unused_mods'), key = '-UNUSED-', size = (47, 1)), sg.Button(tr.getTrans('remove'), key = '-REMOVE-', size = (47, 1))],
 					[sg.Text('')],
 					[sg.Button(tr.getTrans('exit'), key = '-EXIT-', size = (14, 1))]
 				]
@@ -189,6 +240,8 @@ def guiImportMods(updateSGs = True):
 			window['-MODS_INST-'].update(getAllMods())
 		elif event == '-MOD_PATH-':
 			window['-MODS-'].update(values = getMods(values['-MOD_PATH-']))
+		elif event == '-UNUSED-':
+			markUnusedMods(window)
 	window.close()
 	return
 
@@ -321,7 +374,7 @@ def importSGC(path, title):
 			os.mkdir(se.getSettings('fs_game_data_path') + os.sep + new['folder'] + '_Backup')
 		except FileExistsError:
 			pass
-		TinyDB(se.games_json).insert({"name": title, "folder": new['folder'], "desc": new['desc'], "map": new['map'], "mods": new['mods']})
+		TinyDB(se.games_json).insert({"name": title, "folder": new['folder'], "desc": new['desc'], "map": new['map'], "mods": new['mods'], "imported": fdata})
 
 def guiImportSG(path = '', rem = False, overwrite = False):
 	ret = True
